fixed placeholder but flag from is weird
